LinkShield/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-whois==0.8.0
tldextract==5.1.1
joblib==1.3.2
xgboost==2.0.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
